+++ udata/core/dataservices/permissions.py
@@ -1,0 +1,7 @@
+from udata.core.dataset.permissions import (
+    OwnablePermission
+)
+
+class DataserviceEditPermission(OwnablePermission):
+    '''Permissions to edit a Dataservice'''
+    pass